+++ changelog.md
@@ -17,6 +17,7 @@ v1.6 (2022-xx-xx):
 * Improve "ai_catchup" behavior (which also now works in singleplayer)
 * Fix attempting to target things too far down in goop
 * Fix forgetting how to use shield abilities while holding a Fire Flower
+* Improve use of Armageddon Shield nuke to also consider target health and fret status
 
 v1.5 (2022-01-26):
 ------------------